requirements.txt
asgiref==3.7.2
certifi==2026.1.4
charset-normalizer==3.4.4
Django==3.2.25
djangorestframework==3.15.1
idna==3.10
PyMySQL==1.1.1
python-dotenv==0.21.1
pytz==2025.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
